Maven plugin for eclipse changed to only recreate build.properties for binary plugins
git-svn-id: https://overture.svn.sourceforge.net/svnroot/overture/trunk@847 2b2d4d09-c93c-414a-ad04-6e973f5c1cff

+++ tools/maven-pst/maven-psteclipse-plugin/src/main/java/org/codehaus/pst/plugin/ManifestGenerator.java
@@ -157,9 +157,15 @@ public class ManifestGenerator extends AbstractMojoHelper implements
 		File buildPropetiesFile = new File(destinationDirectory,
 				EclipseConstants.BUILD_PROPERTIES);
 
-		if(buildPropetiesFile.exists())
-			buildPropetiesFile.delete();
-		
+		if (buildPropetiesFile.exists())
+		{
+			if (project.getPackaging().equals(
+					EclipseConstants.PACKING_SOURCE_PLUGIN))
+				return;
+			else
+				buildPropetiesFile.delete();
+		}
+
 		if (!buildPropetiesFile.exists())
 		{
 			PrintWriter out = null;
@@ -172,19 +178,20 @@ public class ManifestGenerator extends AbstractMojoHelper implements
 
 				outputFileReader = new FileWriter(buildPropetiesFile);
 
-			BufferedWriter	outputStream = new BufferedWriter(outputFileReader);
+				BufferedWriter outputStream = new BufferedWriter(
+						outputFileReader);
 
-				 out
-				   = new PrintWriter(outputStream);
-				
-				String [] tmp = EclipseConstants.BUILD_PROPERTIES_CONTENT.replace('\n', '#').split("#");
-				
-				for(int i = 0; i < tmp.length ; i++)
+				out = new PrintWriter(outputStream);
+
+				String[] tmp = EclipseConstants.BUILD_PROPERTIES_CONTENT.replace(
+						'\n',
+						'#').split("#");
+
+				for (int i = 0; i < tmp.length; i++)
 				{
-//					getLog().info(tmp[i]);
-				out.println(tmp[i]);
+					// getLog().info(tmp[i]);
+					out.println(tmp[i]);
 				}
-				
 
 			} catch (IOException e)
 			{
